backend/requirements.txt
asgiref==3.11.1
Django==5.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
inflection==0.5.1
Markdown==3.10.2
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
stripe>=5.0.0
uritemplate==4.2.0
